% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedcompendium-package.R
\docType{package}
\name{gedcompendium-package}
\alias{gedcompendium}
\alias{gedcompendium-package}
\title{gedcompendium: Load `Tidyged` and Sister Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Load the ecosystem of packages around the `Tidyged` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jl5000/gedcompendium}
}

}
\author{
\strong{Maintainer}: Jamie Lendrum \email{jalendrum@gmail.com}

}
\keyword{internal}
